Created AP struct
This struct is filled with data received from AP.
ATM, we have the same fields in the `telemetry_t` struct.
This is because the script that converts the telem struct to python only
handles primitives,
but in the future it could perhaps handle internal structs too.
Then telemetry_t should contain `ap_t` instead...